Pass a skillset into Song::HighestMSDOFSkillset directly
Earlier, the skillset was needlessly being cast into an int and then back into a skillset.

+++ src/Etterna/Models/Songs/SongUtil.cpp
@@ -382,11 +382,9 @@ CompareSongPointersByMSD(const Song* pSong1, const Song* pSong2, Skillset ss)
 {
 	// Prefer transliterations to full titles
 	float msd1 = pSong1->HighestMSDOfSkillset(
-	  static_cast<int>(ss),
-	  GAMESTATE->m_SongOptions.Get(ModsLevel_Current).m_fMusicRate);
+	  ss, GAMESTATE->m_SongOptions.Get(ModsLevel_Current).m_fMusicRate);
 	float msd2 = pSong2->HighestMSDOfSkillset(
-	  static_cast<int>(ss),
-	  GAMESTATE->m_SongOptions.Get(ModsLevel_Current).m_fMusicRate);
+	  ss, GAMESTATE->m_SongOptions.Get(ModsLevel_Current).m_fMusicRate);
 
 	if (msd1 < msd2)
 		return true;
